Enhanced script that toggles csv files to be loaded into the database.  The script will now complain and exit if the necessary files are not present to from compressed oplan to extended oplan and back.

+++ csmart/data/database/scripts/mysql/toggle_ext_oplan.sh
@@ -30,23 +30,50 @@ fi
 
 
 if [ "$1" = "extended" ]; then
+    echo "Moving extended oplan versions of csv files to loadable versions."
+
+    if [ ! -e "$COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_ext.csv" ]; then
+        echo "Cannot find $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_ext.csv"
+        exit
+    fi
+    if [ ! -e "$COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_ext.csv" ]; then
+        echo "Cannot find $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_ext.csv"
+        exit
+    fi 
+    if [ ! -e "$COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_ext.csv" ]; then
+        echo "Cannot find $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_ext.csv"
+        exit
+    fi 
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_orig.csv
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_ext.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr.csv
-
+    
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_orig.csv
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_ext.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage.csv
-
+    
     mv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a.csv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_orig.csv
     mv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_ext.csv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a.csv
-
+    echo "Database is now in extended oplan timeframe format."
 else
+    echo "Moving compressed oplan versions of csv files to loadable versions."
+    if [ ! -e "$COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_orig.csv" ]; then
+        echo "Cannot find $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_orig.csv"
+        exit
+    fi 
+    if [ ! -e "$COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_orig.csv" ]; then
+        echo "Cannot find $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_orig.csv"
+        exit
+    fi 
+    if [ ! -e "$COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_orig.csv" ]; then
+        echo "Cannot find $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_orig.csv"
+        exit
+    fi 
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_ext.csv
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr_orig.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_agent_attr.csv
-
+    
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_ext.csv
     mv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage_orig.csv $COUGAAR_INSTALL_PATH/csmart/data/database/csv/oplan_stage.csv
-
+    
     mv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a.csv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_ext.csv
     mv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a_orig.csv $COUGAAR_INSTALL_PATH/fcsua/data/database/csv/mission_period_3a.csv
-
+    echo "Database is now in compressed oplan timeframe format."
 fi
